requirements.txt
numpy~=1.24.4
open3d~=0.17.0
pyquaternion~=0.9.9
opencv-python~=4.8.1.78
vg~=2.0.0
trimesh~=3.21.7
scikit-spatial
PyYAML~=6.0.1
Pillow~=10.1.0
scikit-image~=0.21.0